Add a couple of new overarching goal related components. Does not yet send the goal to the backend.
